backend/pyproject.toml
[tool.poetry]
name = "fastapi auth routes"
version = "0.1.0"
description = ""
authors = ["hojiakbar2806 <hojiakbarnasriddinov2006@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
aiosqlite = "^0.20.0"
asyncpg = "^0.30.0"
pydantic-settings = "^2.7.0"
python-dotenv = "^1.0.1"
pyjwt = "^2.10.1"
bcrypt = "^4.2.1"
python-slugify = "^8.0.4"
python-multipart = "^0.0.20"
alembic = "^1.14.0"
cryptography = "^44.0.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
